classfmt: fix "Implicit narrowing conversion" warnings
ClassFileStruct.u4At(int) is supposed to return an unsigned integer. The
return value is currently always used as primitive integer (signed) to
compute readOffset inside the class file (which will then passed as
argument to the method again). So any negative value would trigger an
ArrayIndexOutOfBoundsException when used as index into the byte array.

The intermediate usage of primitive long did not help and only triggered
warnings at the call sites.

https://github.com/eclipse-jdt/eclipse.jdt.core/security/code-scanning/18
ff

+++ org.eclipse.jdt.core/model/org/eclipse/jdt/internal/core/util/ClassFileReader.java
@@ -85,7 +85,7 @@ public class ClassFileReader extends ClassFileStruct implements IClassFileReader
 		int constantPoolCount;
 		int[] constantPoolOffsets;
 		try {
-			this.magicNumber = (int) u4At(classFileBytes, 0, 0);
+			this.magicNumber = u4At(classFileBytes, 0, 0);
 			if (this.magicNumber != 0xCAFEBABE) {
 				throw new ClassFormatException(ClassFormatException.INVALID_MAGIC_NUMBER);
 			}
@@ -237,7 +237,7 @@ public class ClassFileReader extends ClassFileStruct implements IClassFileReader
 						readOffset += 8;
 						if (attributeCountForField != 0) {
 							for (int j = 0; j < attributeCountForField; j++) {
-								int attributeLength = (int) u4At(classFileBytes, 2, readOffset);
+								int attributeLength = u4At(classFileBytes, 2, readOffset);
 								readOffset += (6 + attributeLength);
 							}
 						}
@@ -263,7 +263,7 @@ public class ClassFileReader extends ClassFileStruct implements IClassFileReader
 						readOffset += 8;
 						if (attributeCountForMethod != 0) {
 							for (int j = 0; j < attributeCountForMethod; j++) {
-								int attributeLength = (int) u4At(classFileBytes, 2, readOffset);
+								int attributeLength = u4At(classFileBytes, 2, readOffset);
 								readOffset += (6 + attributeLength);
 							}
 						}
@@ -323,7 +323,7 @@ public class ClassFileReader extends ClassFileStruct implements IClassFileReader
 						} else {
 							this.attributes[attributesIndex++] = new ClassFileAttribute(classFileBytes, this.constantPool, readOffset);
 						}
-						long tmp = u4At(classFileBytes, readOffset + 2, 0);
+						int tmp = u4At(classFileBytes, readOffset + 2, 0);
 						readOffset += (6 + tmp);
 					}
 				} else {
